triVerts are ivec, not vec

+++ src/libged/facetize/subprocess/util.cpp
@@ -133,7 +133,7 @@ bot_is_manifold(struct rt_bot_internal *bot)
     for (size_t j = 0; j < bot->num_vertices ; j++)
 	bot_mesh.vertPos.push_back(glm::vec3(bot->vertices[3*j], bot->vertices[3*j+1], bot->vertices[3*j+2]));
     for (size_t j = 0; j < bot->num_faces; j++)
-	bot_mesh.triVerts.push_back(glm::vec3(bot->faces[3*j], bot->faces[3*j+1], bot->faces[3*j+2]));
+	bot_mesh.triVerts.push_back(glm::ivec3(bot->faces[3*j], bot->faces[3*j+1], bot->faces[3*j+2]));
     bot_manifold = manifold::Manifold(bot_mesh);
     if (bot_manifold.Status() != manifold::Manifold::Error::NoError)
 	return false;
